requirements.txt
streamlit>=1.32.0
yfinance>=0.2.38
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
textblob>=0.18.0
plotly>=5.20.0
feedparser>=6.0.11
requests>=2.31.0
openai>=1.30.0
python-dotenv>=1.0.0
